change to general coordinate fetching
This is a major and probably the last update of the coordinate fetching function. I changed the code in a way that if you enter the id of an airport or NAVAID, the program checks if it is 4 characters long (ICAO) or 5 characters long (NAVAID), and based on this, it will open the necessary file, airports_short.csv or fixes_short.csv. This saved many lines of code, and brought the program to a simpler level, saving memory and other resources.